Split out PE format handling to a separate file
In preparation for adding support for signing kernel modules,
move the base logic for signing PE files into its own source file.
No behavioural changes made.
